fix: call until event callbacks on resume
When calling `Environment.run(until=event)`, the `until` event's
callbacks list is populated with a callback that raises the
`StopSimulation` exception.

When no other events or processes depended on (waited on) the `until`
event, resuming the simulation by re-calling `Environment.run()` worked
fine.

However, if the `until` event had any dependencies within the
simulation, e.g. a process waiting on it, then upon resuming simulation
that `until` event's callbacks would not be called; that event and its
callbacks were lost on the stack of `Environment.step()` when the
`StopSimulation` exception ended simulation during that event's step.

So any simulation that depended on behaviors happening in response to
that `until` event would, upon resumption of the simulation, behave as
if that event had never happened. Which is a bug.

With this change, a user-provided `until` event will be rescheduled such
that when the simulation is resumed, it will be processed first and its
callbacks (minus `StopSimulation.callback`) called. Thus any simulation
behaviors dependent on that event will happen.

N.B. another possible approach to this problem would be to defer raising
`StopSimulation` until after the `until` event's callbacks were called.
I.e. complete that event's step and then terminate the simulation.

That approach was not taken because the `Environment.run()` docstring
indicates "stepping until this [`until`] event has been triggered". And
the `Event.triggered` docstring indicates that 'triggered' means that
"[an event's] callbacks are about to be invoked". This is distinct from
'processed' which means "[an event's] callbacks have been invoked".

As implemented, the simulation will stop at the same "event triggered"
point. There will be no user-observable difference in the stopped
simulation state. The only user-observable difference in behavior will
be after simulation resumes, due to the `until` event's callbacks being
processed correctly.

Fixes: #168